chore: merge data-clearer with content-management project
