--- contentful/contentful-data-clearer/.env.example
@@ -1,3 +1,0 @@
-CONTENTFUL_MASTER_KEY=ABCD
-CONTENTFUL_SPACE_ID=EFGH
-CONTENTFUL_ENVIRONMENT=dev
