--- contentful/contentful-data-clearer/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "contentful-data-clearer",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "contentful-management": "^11.25.3",
-    "dotenv": "^16.4.5"
-  }
-}
