--- contentful/contentful-data-clearer/types.mjs
@@ -1,8 +1,0 @@
-/**
- * A number, or a string containing a number.
- * @typedef {(Object)} ContentfulEnvironmentOptions
- * @property {string} spaceId - The Contentful space ID
- * @property {string} environmentId - The environment name of the Contentful space
- */
-
-export { };
